Change cxx document and tests

+++ modules/voevodin-andrew-color-converter/src/Color_Converter.cxx
@@ -100,7 +100,7 @@ vector<double> color_converter::XYZtoHSV(const vector<double> _XYZ) const {
         HSV[1] = 1 - min / max;
     HSV[2] = max;
 
-    HSV[0] = static_cast<int>(HSV[0] + 0.5);
+    HSV[0] = (HSV[0] + 0.5);
     return HSV;
 }
 
@@ -129,9 +129,9 @@ vector<int> color_converter::XYZtoLAB(const vector<double> _XYZ) const {
     LAB1[2] = (200.0 * (var_Y - var_Z));
     for (unsigned int i = 0; i < LAB.size(); i++) {
         if (LAB1[i] < 0) {
-           LAB[i] = static_cast<int>(LAB1[i] - 0.9);
+           LAB[i] = (LAB1[i] - 0.9);
         } else {
-            LAB[i] = static_cast<int>(LAB1[i] + 0.9);
+            LAB[i] = (LAB1[i] + 0.9);
         }
     }
     return LAB;
@@ -175,7 +175,7 @@ vector<double> color_converter::HSVtoXYZ(const vector<double> HSV) const {
     if ((HSV[0] < 300.0) && (HSV[0] >= 240.0)) RGB = { x, 0.0, c };
     if ((HSV[0] <= 360.0) && (HSV[0] >= 300.0)) RGB = { c, 0.0, x };
     for (unsigned int i = 0; i < RGB.size(); i++)
-        RGB1[i] = static_cast<int>((RGB[i] + m)*255.0 + 0.9);
+        RGB1[i] = ((RGB[i] + m)*255.0 + 0.9);
     return RGBToXYZ(RGB1);
 }
